Migrate checkAccess & useAccessReview
