+++ frontend/public/components/utils/dropdown.jsx
@@ -5,9 +5,8 @@ import * as PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { useTranslation, withTranslation } from 'react-i18next';
 import { CaretDownIcon, MinusCircleIcon, PlusCircleIcon, StarIcon } from '@patternfly/react-icons';
-import { useSafetyFirst } from '@console/dynamic-plugin-sdk';
+import { impersonateStateToProps, useSafetyFirst } from '@console/dynamic-plugin-sdk';
 import { useUserSettingsCompatibility } from '@console/shared';
-import { impersonateStateToProps } from '@console/dynamic-plugin-sdk';
 
 import { checkAccess } from './rbac';
 import { history } from './router';
